fun with yb-not-pg type strictness

+++ harmony/harmonytask/task_type_handler.go
@@ -156,7 +156,7 @@ func (h *taskTypeHandler) considerWork(from string, ids []TaskID) (workAccepted 
 			FROM harmony_task t
 			JOIN unnest($2::bigint[]) AS x(id) ON x.id = t.id
 			WHERE t.owner_id IS NULL
-			ORDER BY array_position($2,t.id)
+			ORDER BY array_position($2, t.id::bigint)
 			LIMIT $3
 			FOR UPDATE SKIP LOCKED
 		)
